+++ CsharpExtras/Api/ICsharpExtrasApi.cs
@@ -61,8 +61,8 @@ namespace CsharpExtras.Api
             (TObj obj, Func<TObj, TBefore> beforeGetter, Func<TObj, TAfter> afterGetter, Func<TBefore, TAfter, TEvent> eventGenerator);
         ICurryDictionary<TKeyOuter, TValOuter> NewGenericCurryDictionaryWrapper<TKeyInner, TKeyOuter, TValInner, TValOuter>
             (ICurryDictionary<TKeyInner, TValInner> backingDictionary,
-            Func<TKeyOuter, TKeyInner> keyInTransform,
-            Func<TKeyInner, TKeyOuter> keyOutTransform,
+            Func<TKeyOuter, int, TKeyInner> keyInTransform,
+            Func<TKeyInner, int, TKeyOuter> keyOutTransform,
             Func<TValOuter, TValInner> valInTransform,
             Func<TValInner, TValOuter> valOutTransform);
     }