+++ CsharpExtras/Map/Dictionary/Curry/Wrapper/GenericCurryDictionaryWrapperImpl.cs
@@ -1,4 +1,5 @@
-﻿using CsharpExtras.Extensions;
+﻿using CsharpExtras.Api;
+using CsharpExtras.Extensions;
 using CsharpExtras.Extensions.Helper.Dictionary;
 using CsharpExtras.ValidatedType.Numeric.Integer;
 using System;
@@ -11,23 +12,33 @@ namespace CsharpExtras.Map.Dictionary.Curry.Wrapper
     internal class GenericCurryDictionaryWrapperImpl<TKeyInner, TKeyOuter, TValInner, TValOuter>
         : ICurryDictionary<TKeyOuter, TValOuter>
     {
-        ICurryDictionary<TKeyInner, TValInner> _backingDictionary;
-        Func<TKeyOuter, TKeyInner> _keyInTransform;
-        Func<TKeyInner, TKeyOuter> _keyOutTransform;
-        Func<TValOuter, TValInner> _valInTransform;
-        Func<TValInner, TValOuter> _valOutTransform;
+        private ICurryDictionary<TKeyInner, TValInner> _backingDictionary;
+        private Func<TKeyOuter, int, TKeyInner> _keyInTransform;
+        private Func<TKeyInner, int, TKeyOuter> _keyOutTransform;
+        private Func<TValOuter, TValInner> _valInTransform;
+        private Func<TValInner, TValOuter> _valOutTransform;
+        private readonly ICsharpExtrasApi _api;
+
+        private readonly ILazyFunctionMap<int, Func<Func<TKeyInner, int, TKeyOuter>, Func<TKeyInner, int, TKeyOuter>>> _arityInnerOuterCurryMap;
+        private readonly ILazyFunctionMap<int, Func<Func<TKeyOuter, int, TKeyInner>, Func<TKeyOuter, int, TKeyInner>>> _arityOuterInnerCurryMap;
 
         public GenericCurryDictionaryWrapperImpl(ICurryDictionary<TKeyInner, TValInner> backingDictionary,
-            Func<TKeyOuter, TKeyInner> keyInTransform,
-            Func<TKeyInner, TKeyOuter> keyOutTransform,
+            Func<TKeyOuter, int, TKeyInner> keyInTransform,
+            Func<TKeyInner, int, TKeyOuter> keyOutTransform,
             Func<TValOuter, TValInner> valInTransform,
-            Func<TValInner, TValOuter> valOutTransform)
+            Func<TValInner, TValOuter> valOutTransform,
+            ICsharpExtrasApi api)
         {
             _backingDictionary = backingDictionary ?? throw new ArgumentNullException(nameof(backingDictionary));
             _keyOutTransform = keyOutTransform ?? throw new ArgumentNullException(nameof(keyOutTransform));
             _keyInTransform = keyInTransform ?? throw new ArgumentNullException(nameof(keyInTransform));
             _valOutTransform = valOutTransform ?? throw new ArgumentNullException(nameof(valOutTransform));
+            _api = api;
             _valInTransform = valInTransform ?? throw new ArgumentNullException(nameof(valInTransform));
+            _arityInnerOuterCurryMap = api.NewLazyFunctionMap<int, Func<Func<TKeyInner, int, TKeyOuter>, Func<TKeyInner, int, TKeyOuter>>>
+                (GetCurriedKeyTransformFunctionMap<TKeyInner, TKeyOuter>);
+            _arityOuterInnerCurryMap = api.NewLazyFunctionMap<int, Func<Func<TKeyOuter, int, TKeyInner>, Func<TKeyOuter, int, TKeyInner>>>
+                (GetCurriedKeyTransformFunctionMap<TKeyOuter, TKeyInner>);
         }
 
         public TValOuter this[params TKeyOuter[] keyTuple] => _valOutTransform(_backingDictionary[keyTuple.Map(_keyInTransform)]);
@@ -140,31 +151,55 @@ namespace CsharpExtras.Map.Dictionary.Curry.Wrapper
 
         public void UpdateFirstKeyInTuples(Func<TKeyOuter, TKeyOuter> keyInjection)
         {
-            _backingDictionary.UpdateFirstKeyInTuples(ReverseWrapKeyEndoFunction(keyInjection));
+            _backingDictionary.UpdateFirstKeyInTuples(ReverseWrapKeyEndoFunction(keyInjection, 0));
         }
 
         public void UpdateKeys(Func<TKeyOuter, TKeyOuter> keyInjection, NonnegativeInteger prefixArity)
         {
-            _backingDictionary.UpdateKeys(ReverseWrapKeyEndoFunction(keyInjection), prefixArity);
+            _backingDictionary.UpdateKeys(ReverseWrapKeyEndoFunction(keyInjection, prefixArity), prefixArity);
         }
 
         private GenericCurryDictionaryWrapperImpl<TKeyInner, TKeyOuter, TValInner, TValOuter>
             Wrap(ICurryDictionary<TKeyInner, TValInner> dict)
         {
+            Func<Func<TKeyInner, int, TKeyOuter>, Func<TKeyInner, int, TKeyOuter>>
+                keyOutTransformFunction = _arityInnerOuterCurryMap[dict.Arity];
+            Func<Func<TKeyOuter, int, TKeyInner>, Func<TKeyOuter, int, TKeyInner>>
+                keyInTransformFunction = _arityOuterInnerCurryMap[dict.Arity];
             return new GenericCurryDictionaryWrapperImpl<TKeyInner, TKeyOuter, TValInner, TValOuter>
-                (dict, _keyInTransform, _keyOutTransform, _valInTransform, _valOutTransform);
+                (dict, keyInTransformFunction(_keyInTransform), keyOutTransformFunction(_keyOutTransform), _valInTransform, _valOutTransform, _api);
         }
 
         private GenericCurryDictionaryWrapperImpl<TKeyOuter, TKeyInner, TValOuter, TValInner>
-            ReverseWrap(ICurryDictionary<TKeyOuter, TValOuter> other)
+            ReverseWrap(ICurryDictionary<TKeyOuter, TValOuter> dict)
         {
+            Func<Func<TKeyInner, int, TKeyOuter>, Func<TKeyInner, int, TKeyOuter>>
+                keyOutTransformFunction = _arityInnerOuterCurryMap[dict.Arity];
+            Func<Func<TKeyOuter, int, TKeyInner>, Func<TKeyOuter, int, TKeyInner>>
+                keyInTransformFunction = _arityOuterInnerCurryMap[dict.Arity];
             return new GenericCurryDictionaryWrapperImpl<TKeyOuter, TKeyInner, TValOuter, TValInner>
-                (other, _keyOutTransform, _keyInTransform, _valOutTransform, _valInTransform);
+                (dict, keyOutTransformFunction(_keyOutTransform), keyInTransformFunction(_keyInTransform), _valOutTransform, _valInTransform, _api);
         }
 
-        private Func<TKeyInner, TKeyInner> ReverseWrapKeyEndoFunction(Func<TKeyOuter, TKeyOuter> keyInjection)
+        private Func<Func<TKey1, int, TKey2>, Func<TKey1, int, TKey2>> GetCurriedKeyTransformFunctionMap<TKey1, TKey2>(int arity)
         {
-            return k => _keyInTransform(keyInjection(_keyOutTransform(k)));
+            return f => GetCurriedKeyTransform(f, arity);
+        }
+
+        private Func<TKey1, int, TKey2> GetCurriedKeyTransform<TKey1, TKey2>(Func<TKey1, int, TKey2> transform, int arity)
+        {
+            if(arity > Arity)
+            {
+                throw new ArgumentException($"Cannot curry function for arity of {arity} as it exceeds this dictionary's arity of {Arity}");
+            }
+            NonnegativeInteger offset = (NonnegativeInteger)(Arity - arity);
+            return (k, i) => transform(k, i + offset);
+        }
+
+        private Func<TKeyInner, TKeyInner> ReverseWrapKeyEndoFunction(Func<TKeyOuter, TKeyOuter> keyInjection,
+            int indexZeroBased)
+        {
+            return k => _keyInTransform(keyInjection(_keyOutTransform(k, indexZeroBased)), indexZeroBased);
         }
     }
 }